requirements.txt
Django==2.2.13
djangorestframework==3.8.2
pytz==2020.1
sqlparse==0.3.1
